Fix django 1.9 warning and drop support django < 1.7
Fixes a warning that happens when running with Django 1.9:
RemovedInDjango110Warning: django.conf.urls.patterns() is deprecated and will be removed in Django 1.10. Update your urlpatterns to be a list of django.conf.urls.url() instances instead.

Drop support of django < 1.7
Remove the ImportError catching because it was a hack for django < 1.7